+++ CHANGELOG.md
@@ -8,4 +8,8 @@
 
 ## 0.0.3
 
-* Fix pagesView behavior
+* Fix pagesView behavior.
+
+## 0.0.4
+
+* ItemsPerPage Component added.